Logout button and Recovery pop window added
Logout button is located in the user's setting screen. It currently has
no effect on the actual firebase authentication will need to read more
into that. The app has the ability to skip past the login if there's a
logged in user but we have that disabled for now. A logout button was
added in so that the user can return to the login screen.

The recovery pop up window now comes up on recover account need to add
transparency.